+++ lib/appkit/configuration.rb
@@ -3,7 +3,7 @@ module Appkit
     DEFAULT_ICONS = %w[/icon.svg /icon-192.png /icon-512.png /icon-mask-512.png].freeze
 
     attr_accessor :app_name, :email_attribute, :user_scope, :user_class, :first_run,
-                  :icons, :sw_extra_cache_paths, :brand_color
+                  :icons, :sw_extra_cache_paths, :brand_color, :timezone_attribute
 
     def initialize
       @email_attribute = :email
@@ -12,6 +12,7 @@ module Appkit
       @first_run = ->(user_params) { user_class.call.create!(user_params.merge(role: :administrator)) }
       @icons = DEFAULT_ICONS
       @sw_extra_cache_paths = []
+      @timezone_attribute = nil
     end
   end
 end